prplmesh: do not overwrite uci settings
Putting all the uci settings directly under /etc/config means by
default they are overwritten every time prplmesh is reinstalled.

Use a uci-default script instead, to only populate the uci settings
with default value if there is no prplmesh setting at all yet.

The settings are moved as is to the new script, except that the syntax
is different since it's using uci commands.

While we're at it:

- Merge 99-prplmesh-update-config.sh into the new script.

- The config directory is now empty, so remove it from the Makefile.

- Place the new scripts under /etc/uci-defaults, instead of directly
  under 'files'.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>