prplmesh-dwpal: fix bridge package dependency
prplMesh currently depends on `bridge` package which is patched in order
to provide UGW platform specific bridge features which seems to be
needed only by prplmesh-dwpal variant and not needed by other variants.

So make it clear and move that UGW patched `bridge` dependency to
prplmesh-dwpal variant and while at it use new package `bridge-ugw`
which was made available to properly handle the package dependencies.

The problem is, that `bridge` package is available from OpenWrt packages
feed as well, which obviously doesn't contain the UGW features. Resolve
this ambiguity by explicitly demanding proper `bridge-ugw` package.

Signed-off-by: Petr Štetiar <ynezz@true.cz>