prplmesh: init.d: add timeout to wlan ready poll
https://jira.prplfoundation.org/browse/PPM-312

Currently the init script waits in an endless loop until all wireless
interfaces are ready before starting prplMesh.

It occurs that after flashing a new image to a Gl.iNet B1300 device,
both radios are disabled by default so the loop waits indefinitely and
it looks like the init script is hanged out.

To avoid this situation, sit on the loop until the condition is met or a
timeout elapses.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>