libtrace: use --without-lzma explicitly
Currently libtrace is configured without any explicit
'{with,without}-lzma' option, which means it tries to detect lzma
support by itself at configure time.

However, it can happen that lzma is detected as available at configure
time, even though it is not available later in the build as a compiled
library. When this happens, the build fails with:
Package libtrace is missing dependencies for the following libraries: liblzma.so.5

To fix this, explicitly use '--without-lzma', since liblzma is anyway
not in the DEPENDS.

While at it, use tabs instead of spaces for the other configure
arguments, to keep the file consistent (and also consistent with what
is done in other packages).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>