prplmesh: init.d: don't change bridge settings
It's not the responsibility of the prplMesh init script to configure
the interfaces names that are part of the lan network.

Besides, the interface names that are used may not even exist on the
device that is being configured, which means the init script would
just render the device unusable.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>