prplmesh: init.d: fix setting certification_mode
__uci_set_certification_mode__ was adding a specific IP for the
interfaces that are part of the 'lan' network.

It is not always desirable to use this IP, so remove it.

There is something important that __uci_set_certification_mode__()
didn't do though: setting the certification_mode itself. Since it's
what we would expect to happen after calling the script with
"certification_mode", add it.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>