prplmesh: fix libamxb dependency not being included
With the ifdef construct, libamxb was not added to the dependencies
because the config symbol is undefined when the dependencies are
resolved.

To fix it, do like it's done for TARGET_intel_mips: use the
"PACKAGE_libamxb:package" construct. Note that in this case we need to
add a dependency on "libamxb" itself if it's already selected, to make
sure it's built before prplMesh.

Also add PACKAGE_libamxb to PKG_CONFIG_DEPENDS to make sure the
prplMesh package is reconfigured if the value of the config symbol
changes (because some cmake options depend on it).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>