prplmesh: bump to 1.7.0
When gitlab creates a tarball for a tag, it normally creates a tarball
with a top-level directory named "<pkg>-<tag>-<sha1>" - except if we
request for a tarball name "<pkg>-<tag>.tar.gz", then it will use just
"<pkg>-<tag>" as the top-level directory. Since we would prefer the
latter one, make sure that we use that tarball name. Case is significant
here: the OpenWRT package is called "prplmesh" while the upstream
package is "prplMesh".

Remove 0001-bpl-arp-fix-libbridge-dependency.patch because that patch
has already been applied in prplMesh.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>