luci-mod-simple: add prplMesh
For prplMesh to be usable with only WebUI access, there has to be a
way to configure it as either controller or agent.

Add a page for prplMesh, with a radio button to configure the
mode. All of it is taken from the firewall page, and adapted for
prplMesh.

Note that changing the prplMesh mode requires to set multiple UCI
settings (some of them are only useful for other platforms). By
simplicity querying the current mode is done by looking only at
'management_mode', instead of going through all of the settings (from
the WebUI, there is anyway only a way to set all of them at once).

Applying the settings requires prplMesh to be restarted after 'uci
commit' has been called. For that, add an rpcd script. In the future
if prplMesh switches to using procd to manage its processes, it will
be possible to remove the rpcd script and replace it with a procd
trigger.

https://jira.prplfoundation.org/browse/PPM-1355

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
(cherry picked from commit 15879446daf45bb3d200f945e96a20faaadebb4d)