prplmesh: uci-defaults: set agent as the default mode
Since (at least with the current prplMesh revision) only one
controller can be in the network, it's easier for users to have
prplMesh running as an agent by default, and to configure only one as
the controller.

Note that some of the options (e.g. wired_backhaul) are unused and
will be removed in the future, but for now they are changed as well to
keep them in sync with the prplMesh init.d file.

https://jira.prplfoundation.org/browse/PPM-1333

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
(cherry picked from commit 971d503e921c80848d9bcbbb6ea7a636bf7d67a9)