prplmesh: uci-defaults: use the first lan interface for wired backhaul
When the prplMesh agent starts, it checks that the wired backhaul
interface is part of the br-lan bridge. For this reason, the wan
ainterface is not a good default choice.

Change the default to use the first interface of `network.lan`. In
case we couldn't get any interface, raise a warning.

While at it, remove the checks for an existing
'prplmesh.config.backhaul_wire_iface', as this script anyway aborts
early if there is an existing configuration.

Closes PPM-1333

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>