luci-mod-simple: prplmesh: add an option to change backhaul_wire_iface
Currently prplMesh supports only one interface to be used for wired
backhaul. This means that the related UCI option has to be set
appropriately, for a prplMesh agent to be able to find the controller.

Add an option in the prplMesh page to select an interface among the
interfaces that are part of `network.lan.ifname`.

Even though the UCI option is located in the same section as the
operating mode, create a new section to visually separate the backhaul
settings from the mode.

The different handlers to save are added back, but only apply to the
backhaul settings (as the mode has it's own way to apply settings).

A note is also added that prplMesh has to be manually restarted after
making changes, as it currently doesn't detect UCI changes
automatically.

Closes PPM-1375

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>