prplMesh: use PKG_REV for PRPLMESH_REVISION
prplMesh used to be built from a git repository, and was back then
using PKG_SOURCE_VERSION to set the version.

Since `fc848dcd516bd779330cf36e0de2b76619101559` however prplMesh is
built from a tarball, so PKG_SOURCE_VERSION is not set. As a
consequence, prplMesh falls back to using git to get the package
revision. When building in-tree, the git command ends up getting the
revision for the prplWrt repository itself, and not for prplMesh.

Fix it by using PKG_REV instead.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>