tr181-rest-api: add new package
This package contains TR181 based REST API powered by lighttpd and using
amx-fcgi module as a backend.

Not adding this new `tr181-rest-api` package to the list of CI compile
tested packages as this package depends on lighttpd which is currently
broken and can't be built properly with SDK. We probably don't need to
CI build this package as `amx-fcgi` package is already build tested and
`tr181-rest-api` package contains only configuration files, there is no
compile testing.

Closes PCF-611
Signed-off-by: Petr Štetiar <ynezz@true.cz>