tr181-rest-api: make authentication configuration fine-grained
Currently the example authentication is configured in a coarse-grained
way, so it demands valid authentication for every URL being handled by
this server, which makes it cumbersome for any 3rd party integration.

So let's make the authentication more fine-grained and lets enable it
only on REST API endpoints provided by this package.

Signed-off-by: Petr Štetiar <ynezz@true.cz>