time-manager: fix libsahtrace linking with gcc 11.2.0
Currently build fails during linking step of gcc 11.2.0 toolchain due to:

 cannot find -lsahtrace

That's happening due to the fact, that libsahtrace is being installed to
STAGINGDIR/lib, but that path is missing in the linker paths as the
order of the linker arguments seems to matter. So fix it by adding
STAGING_LIBDIR into the linker args before any other args.

References: PCF-736
Signed-off-by: Petr Štetiar <ynezz@true.cz>