PROVISORY: ci: import generate SDK package build job
pwhm is currently failing to build with gcc-11 used in OpenWrt 22.03
SDK, so we're not going to build test it with that SDK temporarily until
PPM-2363 is finished. Meanwhile we're simply going to move forward and
disable building of pwhm with that failing SDK.

For that I'm going to temporarily override the CI job to exclude pwhm
package from build if OpenWrt 22.03 SDK is going to be used. This
modification is going to happen in the upcomming commit, so the changes
are in separate commit and easier to review.

References: PCF-736
Signed-off-by: Petr Štetiar <ynezz@true.cz>