PROVISORY: ci: do not build pwhm and mod-whm-dummy with OpenWrt 22.03 SDK
pwhm and mod-whm-dummy are currently failing to build with gcc-11 used
in OpenWrt 22.03 SDK, so we're not going to build test those packages
with that SDK temporarily until PPM-2363 is finished. Meanwhile we're
simply going to move forward and disable building of pwhm and
mod-whm-dummy packages with that upcoming SDK version.

References: PCF-736
Signed-off-by: Petr Štetiar <ynezz@true.cz>