prplmesh: fix recursive dependencies for Ambiorix
Currently we're getting following recursive dependency errors:

 tmp/.config-package.in:2267:error: recursive dependency detected!
 tmp/.config-package.in:2267:    symbol PACKAGE_mod-amxb-ubus is selected by PACKAGE_libamxb
 tmp/.config-package.in:1590:    symbol PACKAGE_libamxb is selected by PACKAGE_mod-amxb-ubus

So lets fix those by introducing PRPLMESH_ENABLE_AMBIORIX_UBUS_BACKEND
config option and while at it, enable it by default as well, since
Ambiorix used to be optional for prplMesh, but we should start
considering it required.

Without Ambiorix, for example pWHM can't work, and also the only API
that's available is the non-standard (and deprecated) BML. To use the
(proper) Northbound API or the pWHM or the new persistence features that
will be developed, Ambiorix is necessary anyway.

Signed-off-by: Petr Štetiar <ynezz@true.cz>