prplmesh: workaround hostapd-openssl related recursive dependecies
It seems, that upstream in commit 009293c52e63 ("build: scripts/config -
update to kconfig-v5.14") has changed handling of recursive dependencies,
which currently results in the following:

 tmp/.config-package.in:29231:error: recursive dependency detected!
 tmp/.config-package.in:29231:  symbol PACKAGE_prplmesh depends on PACKAGE_hostapd-openssl
 tmp/.config-package.in:31240:  symbol PACKAGE_hostapd-openssl is selected by PACKAGE_prplmesh

and unfortunately results in the missing prplmesh package in the `.config`.

This recursive dependencies in package prplmesh seems to be caused by
following dependency chain:

 prplmesh
  |  |
  |  `-> hostapd-openssl <-.
  |                        |
  `----> hostapd-utils ----'

So lets try to workaround it by moving `hostapd-utils` package into
respective target YAML profiles.

Signed-off-by: Petr Štetiar <ynezz@true.cz>