libsqla/libswlc/pwhm: (APP-3656, PR #43) remove prplmesh package errors
Some prplmesh packages rely on host packages that don’t exist in any
of our feeds.  Since we’re not using these packages at the moment,
let’s just kill the dependencies.

(note that the package build will probably still work as long as the
correct libs are on the host)


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn