prplmesh: (APP-3936, PR #56) Refresh Easymesh topology to fetch backhaul info
EasyMesh Background:
The EasyMesh controller relies on a combination of topology discovery, notifications, and query/response exchanges to build and maintain the network topology. These messages are exchanged during various network events—such as when an agent joins or leaves the network, a station connects or disconnects, or when an interface goes up or down. When a change occurs, an agent broadcasts a topology notification to all EasyMesh nodes, prompting the controller to update its view by initiating a topology query/response exchange.

Expected Behaviour:
For the correct data to appear in the Device.WiFi.DataElements.Network.Device.X.MultiAPDevice.Backhaul element, the following sequence should occur:

A Topology Notification is received from the local agent, indicating that the backhaul station is connected.

Subsequently, a Topology Response is received from the extender agent, containing the latest topology information.

Issue Scenario:
In some cases, the Topology Response arrives before the Topology Notification. As a result, the response lacks the necessary details to properly populate the Device.WiFi.DataElements.Network.Device.X.MultiAPDevice.Backhaul element. The Topology Notification is delayed because the prplmesh agent defers sending it until the auto-configuration process is complete.

Fix:
If the backhaul information is missing in the Device.WiFi.DataElements.Network.Device.X.MultiAPDevice.Backhaul element, trigger a topology query to the agent to ensure that the controller receives the complete and correct data.

Approved-by: Sophronia Koilpillai
Approved-by: James Haggerty